reverts transparent styling, adds kml style injection in new file
